+++ docs/Changelog.md
@@ -10,6 +10,7 @@
   - afl-fuzz:
     - memory leak fixes by @kcwu - thanks!
     - some more nits and small memory saves thanks to @kcwu
+    - remove deprecated files from queue/.state
   - frida_mode:
     - fixes for new MacOS + M4 hardware